works but todo: collect result, render forloop, convert forloopraw to seq for frontend
