+++ backend/src/jobs.rs
@@ -5,22 +5,21 @@
  * LICENSE-AGPL for a copy of the license.
  */
 
-use async_recursion::async_recursion;
 use chrono::Duration;
 
 use sql_builder::prelude::*;
 use sqlx::{query_scalar, Postgres, Transaction};
 use std::collections::HashMap;
+use tracing::instrument;
 
-use crate::js_eval::{eval_timeout, EvalCreds};
 use crate::scripts::{get_hub_script_by_path, ScriptLang};
-use crate::users::create_token_for_owner;
+use crate::worker_flow::init_flow_status;
 use crate::{
     audit::{audit_log, ActionKind},
     db::{UserDB, DB},
     error,
     error::Error,
-    flow::{FlowModuleValue, FlowValue, InputTransform},
+    flow::FlowValue,
     schedule::get_schedule_opt,
     scripts::ScriptHash,
     users::{owner_to_token_owner, Authed},
@@ -33,7 +32,7 @@ use axum::{
 };
 use hyper::StatusCode;
 use serde::{Deserialize, Serialize};
-use serde_json::{json, Map, Value};
+use serde_json::{Map, Value};
 use sql_builder::SqlBuilder;
 
 use ulid::Ulid;
@@ -187,7 +186,7 @@ pub async fn run_job_by_path(
     Ok((StatusCode::CREATED, uuid.to_string()))
 }
 
-async fn script_path_to_payload<'c>(
+pub async fn script_path_to_payload<'c>(
     script_path: &str,
     db: &mut Transaction<'c, Postgres>,
     w_id: &String,
@@ -808,7 +807,7 @@ async fn get_job_from_id<'c>(
     Ok((job_option, tx))
 }
 
-async fn get_queued_job<'c>(
+pub async fn get_queued_job<'c>(
     id: Uuid,
     w_id: &str,
     tx: &mut Transaction<'c, Postgres>,
@@ -842,42 +841,6 @@ pub enum JobKind {
     Dependencies,
     Flow,
     FlowPreview,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct FlowStatus {
-    pub step: i32,
-    pub modules: Vec<FlowStatusModule>,
-    pub failure_module: FlowStatusModule,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Iterator {
-    pub index: u8,
-    pub itered: Vec<Value>,
-    pub args: Map<String, Value>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(tag = "type")]
-pub enum FlowStatusModule {
-    WaitingForPriorSteps,
-    WaitingForEvent {
-        event: String,
-    },
-    WaitingForExecutor {
-        job: Uuid,
-    },
-    InProgress {
-        job: Uuid,
-        iterator: Option<Iterator>,
-    },
-    Success {
-        job: Uuid,
-    },
-    Failure {
-        job: Uuid,
-    },
 }
 
 #[derive(sqlx::FromRow)]
@@ -996,6 +959,7 @@ struct PreviewFlow {
     args: Option<Map<String, Value>>,
 }
 
+#[derive(Debug)]
 pub enum JobPayload {
     ScriptHub {
         path: String,
@@ -1016,6 +980,7 @@ pub enum JobPayload {
     },
 }
 
+#[instrument(level = "trace", skip_all)]
 pub async fn push<'c>(
     mut tx: Transaction<'c, Postgres>,
     workspace_id: &str,
@@ -1149,13 +1114,7 @@ pub async fn push<'c>(
         }
     };
 
-    let flow_status = raw_flow.as_ref().map(|f| FlowStatus {
-        step: 0,
-        modules: (0..f.modules.len())
-            .map(|_| FlowStatusModule::WaitingForPriorSteps)
-            .collect(),
-        failure_module: FlowStatusModule::WaitingForPriorSteps,
-    });
+    let flow_status = raw_flow.as_ref().map(init_flow_status);
     let uuid = sqlx::query_scalar!(
         "INSERT INTO queue
             (workspace_id, id, parent_job, created_by, permissioned_as, scheduled_for, 
@@ -1218,7 +1177,8 @@ pub async fn push<'c>(
     Ok((uuid, tx))
 }
 
-pub async fn add_completed_job_error<E: ToString>(
+#[instrument(level = "trace", skip_all)]
+pub async fn add_completed_job_error<E: ToString + std::fmt::Debug>(
     db: &DB,
     queued_job: &QueuedJob,
     logs: String,
@@ -1241,6 +1201,7 @@ pub async fn add_completed_job_error<E: ToString>(
     Ok((a, output_map))
 }
 
+#[instrument(level = "trace", skip_all)]
 pub async fn add_completed_job(
     db: &DB,
     queued_job: &QueuedJob,
@@ -1293,174 +1254,7 @@ pub async fn add_completed_job(
     Ok(queued_job.id)
 }
 
-pub async fn get_step_of_flow_status(db: &DB, id: Uuid) -> error::Result<i32> {
-    let r = sqlx::query_scalar!(
-        "SELECT (flow_status->'step')::integer FROM queue WHERE id = $1",
-        id
-    )
-    .fetch_one(db)
-    .await?
-    .ok_or_else(|| Error::InternalErr(format!("not found step")))?;
-    Ok(r)
-}
-
-pub async fn update_flow_status_in_progress(
-    db: &DB,
-    w_id: &str,
-    flow: Uuid,
-    job_in_progress: Uuid,
-) -> error::Result<()> {
-    let step = get_step_of_flow_status(db, flow).await?;
-    sqlx::query(&format!(
-        "UPDATE queue
-            SET flow_status = jsonb_set(flow_status, '{{modules, {}}}', $1)
-            WHERE id = $2 AND workspace_id = $3",
-        step
-    ))
-    .bind(serde_json::json!(FlowStatusModule::InProgress {
-        job: job_in_progress,
-        iterator: None,
-    }))
-    .bind(flow)
-    .bind(w_id)
-    .execute(db)
-    .await?;
-    Ok(())
-}
-
-#[async_recursion]
-pub async fn update_flow_status_after_job_completion(
-    db: &DB,
-    job: &QueuedJob,
-    success: bool,
-    result: Option<Map<String, Value>>,
-) -> error::Result<()> {
-    tracing::info!("HANDLE FLOW: {job:?} {success} {result:?}");
-
-    let mut tx = db.begin().await?;
-
-    let w_id = &job.workspace_id;
-
-    let flow = job
-        .parent_job
-        .ok_or_else(|| Error::InternalErr(format!("expected parent job")))?;
-
-    let old_status_json = sqlx::query_scalar!(
-        "SELECT flow_status FROM queue WHERE id = $1 AND workspace_id = $2",
-        flow,
-        w_id
-    )
-    .fetch_one(&mut tx)
-    .await?
-    .ok_or_else(|| Error::InternalErr(format!("requiring a previous status")))?;
-
-    let old_status = serde_json::from_value::<FlowStatus>(old_status_json)
-        .ok()
-        .ok_or_else(|| {
-            Error::InternalErr(format!("requiring status to be parsabled as FlowStatus"))
-        })?;
-
-    let last_step = (old_status.step + 1) as usize == old_status.modules.len();
-
-    let (step_counter, new_status) = match &old_status.modules[old_status.step as usize] {
-        module_status @ FlowStatusModule::InProgress {
-            job: _,
-            iterator:
-                Some(Iterator {
-                    index,
-                    itered,
-                    args: _,
-                }),
-        } if (index.to_owned() as usize) < itered.len() - 1 => {
-            (old_status.step, module_status.clone())
-        }
-        _ => {
-            let new_status = if success {
-                FlowStatusModule::Success { job: job.id }
-            } else {
-                FlowStatusModule::Failure { job: job.id }
-            };
-            (old_status.step + 1, new_status)
-        }
-    };
-
-    sqlx::query(&format!(
-        "UPDATE queue
-            SET 
-                flow_status = jsonb_set(jsonb_set(flow_status, '{{modules, {}}}', $1), '{{\"step\"}}', $2)
-            WHERE id = $3",
-        old_status.step,
-    ))
-    .bind(serde_json::json!(new_status))
-    .bind(serde_json::json!(step_counter))
-    .bind(flow)
-    .execute(&mut tx)
-    .await?;
-
-    tracing::info!("UPDATE: {:?}", new_status);
-
-    let flow_job = get_queued_job(flow, w_id, &mut tx)
-        .await?
-        .ok_or_else(|| Error::InternalErr(format!("requiring flow to be in the queue")))?;
-    tx.commit().await?;
-
-    let done = if !success || last_step {
-        add_completed_job(
-            db,
-            &flow_job,
-            success,
-            false,
-            result.clone(),
-            "Flow job completed".to_string(),
-        )
-        .await?;
-        true
-    } else {
-        match handle_flow(&flow_job, db, result.clone()).await {
-            Err(err) => {
-                let _ = add_completed_job_error(
-                    db,
-                    &flow_job,
-                    "Unexpected error during flow chaining:\n".to_string(),
-                    err,
-                )
-                .await;
-                true
-            }
-            Ok(StopEarly::Continue) => false,
-            Ok(stop_early) => {
-                let _ = add_completed_job(
-                    db,
-                    &flow_job,
-                    true,
-                    matches!(stop_early, StopEarly::Skip),
-                    result.clone(),
-                    "Flow job stopped early".to_string(),
-                )
-                .await;
-                true
-            }
-        }
-    };
-
-    if done {
-        postprocess_queued_job(
-            flow_job.schedule_path,
-            flow_job.script_path,
-            &w_id,
-            flow,
-            db,
-        )
-        .await?;
-
-        if flow_job.parent_job.is_some() {
-            return Ok(update_flow_status_after_job_completion(db, &job, success, result).await?);
-        }
-    }
-
-    Ok(())
-}
-
+#[instrument(level = "trace", skip_all)]
 pub async fn postprocess_queued_job(
     schedule_path: Option<String>,
     script_path: Option<String>,
@@ -1473,6 +1267,7 @@ pub async fn postprocess_queued_job(
     Ok(())
 }
 
+#[instrument(level = "trace", skip_all)]
 pub async fn schedule_again_if_scheduled(
     schedule_path: Option<String>,
     script_path: Option<String>,
@@ -1491,304 +1286,6 @@ pub async fn schedule_again_if_scheduled(
         tx.commit().await?;
     }
     Ok(())
-}
-
-pub async fn handle_flow(
-    job: &QueuedJob,
-    db: &sqlx::Pool<sqlx::Postgres>,
-    last_result: Option<Map<String, serde_json::Value>>,
-) -> anyhow::Result<StopEarly> {
-    let value = job
-        .raw_flow
-        .as_ref()
-        .ok_or_else(|| Error::InternalErr(format!("requiring a raw flow value")))?
-        .to_owned();
-    let flow = serde_json::from_value::<FlowValue>(value.to_owned())?;
-    let stop_early =
-        push_next_flow_job(job, flow, job.schedule_path.clone(), db, last_result).await?;
-    Ok(stop_early)
-}
-
-async fn transform_input(
-    flow_args: &Option<serde_json::Value>,
-    last_result: Option<Map<String, serde_json::Value>>,
-    input_transform: &HashMap<String, InputTransform>,
-    workspace: &str,
-    token: &str,
-    steps: Vec<String>,
-) -> anyhow::Result<Option<Map<String, serde_json::Value>>> {
-    let mut mapped = serde_json::Map::new();
-
-    for (key, val) in input_transform.into_iter() {
-        match val {
-            InputTransform::Static { value } => {
-                mapped.insert(key.to_string(), value.to_owned());
-                ()
-            }
-            _ => (),
-        };
-    }
-
-    for (key, val) in input_transform.into_iter() {
-        match val {
-            InputTransform::Static { value: _ } => (),
-            InputTransform::Javascript { expr } => {
-                let previous_result =
-                    serde_json::Value::Object(last_result.clone().unwrap_or_else(|| Map::new()));
-                let flow_input = flow_args.clone().unwrap_or_else(|| json!({}));
-                let v = eval_timeout(
-                    expr.to_string(),
-                    vec![
-                        ("params".to_string(), serde_json::json!(mapped)),
-                        ("previous_result".to_string(), previous_result),
-                        ("flow_input".to_string(), flow_input),
-                    ],
-                    Some(EvalCreds {
-                        workspace: workspace.to_string(),
-                        token: token.to_string(),
-                    }),
-                    steps.clone(),
-                )
-                .await
-                .map_err(|e| {
-                    Error::ExecutionErr(format!(
-                        "Error during isolated evaluation of expression `{expr}`:\n{e}"
-                    ))
-                })?;
-                mapped.insert(key.to_string(), v);
-                ()
-            }
-        }
-    }
-
-    Ok(Some(mapped))
-}
-
-pub enum StopEarly {
-    Skip,
-    Success,
-    Continue,
-}
-async fn push_next_flow_job(
-    job: &QueuedJob,
-    flow: FlowValue,
-    schedule_path: Option<String>,
-    db: &sqlx::Pool<sqlx::Postgres>,
-    last_result: Option<Map<String, serde_json::Value>>,
-) -> anyhow::Result<StopEarly> {
-    let flow_status_json = job
-        .flow_status
-        .as_ref()
-        .ok_or_else(|| Error::InternalErr(format!("not found status for flow job {:?}", job.id)))?;
-    let status = serde_json::from_value::<FlowStatus>(flow_status_json.to_owned())?;
-    let i = status.step as usize;
-
-    if i > 0 {
-        let prev_module = &flow.modules[i - 1];
-        if let Some(expr) = prev_module.stop_after_if_expr.clone() {
-            let result =
-                serde_json::Value::Object(last_result.clone().unwrap_or_else(|| Map::new()));
-            match eval_timeout(expr, [("result".to_string(), result)].into(), None, vec![]).await? {
-                serde_json::Value::Bool(true) => {
-                    if let Some(true) = prev_module.skip_if_stopped {
-                        return Ok(StopEarly::Skip);
-                    } else {
-                        return Ok(StopEarly::Success);
-                    }
-                }
-                a @ _ => Error::ExecutionErr(format!("Expected a boolean value, found: {a:?}")),
-            };
-        }
-    }
-    if flow.modules.len() > i {
-        let module = &flow.modules[i];
-        let mut tx = db.begin().await?;
-        let job_payload = match &module.value {
-            FlowModuleValue::Script {
-                path: script_path,
-                trigger_script: None,
-            } => script_path_to_payload(script_path, &mut tx, &job.workspace_id).await?,
-            FlowModuleValue::RawScript(raw_code) => JobPayload::Code(raw_code.clone()),
-            FlowModuleValue::ForloopFlow { iterator: _, value } => JobPayload::RawFlow {
-                value: *(*value).to_owned(),
-                path: Some(format!(
-                    "{}/{i}",
-                    job.script_path
-                        .as_ref()
-                        .unwrap_or(&"NO_FLOW_PATH".to_owned())
-                )),
-            },
-            a @ _ => {
-                tracing::info!("Unrecognized module values {:?}", a);
-                Err(Error::BadRequest(format!(
-                    "Unrecognized module values {:?}",
-                    a
-                )))?
-            }
-        };
-
-        let token = create_token_for_owner(
-            &db,
-            &job.workspace_id,
-            &job.permissioned_as,
-            crate::users::NewToken {
-                label: Some("transform-input".to_string()),
-                expiration: Some(chrono::Utc::now() + chrono::Duration::seconds(10)),
-            },
-            &job.created_by,
-        )
-        .await?;
-        let mut input_transform = module.input_transform.clone();
-
-        let (forloop_args, forloop_iterator) = match &module.value {
-            FlowModuleValue::ForloopFlow { iterator, .. } => {
-                let (index_forloop, itered, args) = match &status.modules[i] {
-                    FlowStatusModule::WaitingForPriorSteps { .. } => {
-                        let itered = match iterator {
-                            InputTransform::Static { value } => value.clone(),
-                            InputTransform::Javascript { expr } => {
-                                let result = serde_json::Value::Object(
-                                    last_result.clone().unwrap_or_else(|| Map::new()),
-                                );
-                                eval_timeout(
-                                    expr.to_string(),
-                                    [("result".to_string(), result)].into(),
-                                    None,
-                                    vec![],
-                                )
-                                .await?
-                            }
-                        };
-                        input_transform.insert(
-                            "_index".to_string(),
-                            InputTransform::Static {
-                                value: serde_json::Value::Number(serde_json::Number::from(0)),
-                            },
-                        );
-                        input_transform.insert(
-                            "_values".to_string(),
-                            InputTransform::Static {
-                                value: json!(itered),
-                            },
-                        );
-                        input_transform.insert(
-                            "_value".to_string(),
-                            InputTransform::Static {
-                                value: json!(itered[0]),
-                            },
-                        );
-                        match itered {
-                            serde_json::Value::Array(arr) => (0 as u8, arr, None),
-                            a @ _ => Err(Error::BadRequest(format!(
-                                "Expected an array value, found: {:?}",
-                                a
-                            )))?,
-                        }
-                    }
-                    FlowStatusModule::InProgress {
-                        job: _,
-                        iterator:
-                            Some(Iterator {
-                                index,
-                                itered,
-                                args,
-                            }),
-                    } => {
-                        let mut args = args.clone();
-                        args.insert(
-                            "_index".to_string(),
-                            serde_json::Value::Number(serde_json::Number::from(index.to_owned())),
-                        );
-                        args.insert(
-                            "_value".to_string(),
-                            itered[index.to_owned() as usize].clone(),
-                        );
-                        (index.to_owned() + 1, itered.to_owned(), Some(args))
-                    }
-                    a @ _ => Err(Error::BadRequest(format!(
-                        "Unrecognized module status for ForloopFlow {:?}",
-                        a
-                    )))?,
-                };
-                (args, Some((index_forloop, itered)))
-            }
-            _ => (None, None),
-        };
-
-        let args = if forloop_args.is_some() {
-            forloop_args.map(|x| x.to_owned())
-        } else {
-            let steps = status
-                .modules
-                .into_iter()
-                .map(|x| match x {
-                    FlowStatusModule::Success { job } => job.to_string(),
-                    _ => "invalid step status".to_string(),
-                })
-                .collect();
-
-            let transformed = transform_input(
-                &job.args,
-                last_result,
-                &input_transform,
-                &job.workspace_id,
-                &token,
-                steps,
-            )
-            .await?;
-
-            match (&job.args, &module.value) {
-                (Some(Value::Object(m)), FlowModuleValue::ForloopFlow { .. }) => {
-                    transformed.map(|x| {
-                        let mut args = m.to_owned();
-                        args.extend(x);
-                        args
-                    })
-                }
-                _ => transformed,
-            }
-        };
-
-        let (uuid, mut tx) = push(
-            tx,
-            &job.workspace_id,
-            job_payload,
-            args.clone(),
-            &job.created_by,
-            job.permissioned_as.to_owned(),
-            None,
-            schedule_path,
-            Some(job.id),
-            true,
-        )
-        .await?;
-
-        let new_status = if let Some((index, itered)) = forloop_iterator {
-            serde_json::json!(FlowStatusModule::InProgress {
-                job: uuid,
-                iterator: Some(Iterator {
-                    index: index,
-                    itered: itered,
-                    args: args.unwrap_or_else(|| Map::new()),
-                })
-            })
-        } else {
-            serde_json::json!(FlowStatusModule::WaitingForExecutor { job: uuid })
-        };
-        sqlx::query(&format!(
-            "UPDATE queue
-            SET 
-                flow_status = jsonb_set(flow_status, '{{modules, {}}}', $1)
-            WHERE id = $2",
-            i
-        ))
-        .bind(new_status)
-        .bind(job.parent_job)
-        .execute(&mut tx)
-        .await?;
-        tx.commit().await?;
-    }
-    Ok(StopEarly::Continue)
 }
 
 pub async fn pull(db: &DB) -> Result<Option<QueuedJob>, crate::Error> {
@@ -1815,6 +1312,7 @@ pub async fn pull(db: &DB) -> Result<Option<QueuedJob>, crate::Error> {
     Ok(job)
 }
 
+#[instrument(level = "trace", skip_all)]
 pub async fn delete_job(db: &DB, w_id: &str, job_id: Uuid) -> Result<(), crate::Error> {
     let job_removed = sqlx::query_scalar!(
         "DELETE FROM queue WHERE workspace_id = $1 AND id = $2 RETURNING 1",
